Updated pattern matching to find patterns in subdirectories.
The pattern now only supports a filesystem glob.  It is then split by
the forward slash and each part converted into a regular exression using
fnmatch.translate.  Any occurence of "FILEVERSION" will be replaced with
"(?P<version>[0-9\\.])" to match repeating digits or periods and be
reported as the version, only the last segment's version if used
multiple times is used as the version for autonaming.

[*/*.py]
; this will match all python files under any directory in the
; collection.

[dir/file-FILEVERSION.tar.*]
; This will match the any file with the pattern "file-[0-9\.]+\.tar\..*"
; in the directory "dir", and the digits portion will be treated as the
; version to use for any autoname