Added ability to preserve state during verfication of files.
If the state option is specified during a verify pass, then only files
which have not been succesfully checksum verified will have their
checksums verified again and added to the state file.  This can be
performed to have a check of many large files performed over multiple
passes. CTRL-C can be used to abort a check/verify and save the state
file